langgraph_server_demo/pyproject.toml
[project]
name = "langgraph-server-demo"
version = "0.1.0"
description = "LangGraph Server demo with deepagents"
requires-python = ">=3.11,<3.14"
dependencies = [
    "deepagents",
    "langgraph-cli[inmem]>=0.1.56",
    "langgraph-checkpoint-sqlite",
    "httpx",  # for client_example.py
]

[dependency-groups]
dev = []

[tool.uv.sources]
deepagents = { path = "../libs/deepagents", editable = true }
